requirements.txt
absl-py==2.1.0
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
datasets==2.19.1
debugpy==1.8.1
decorator==5.1.1
dill==0.3.8
evaluate==0.4.2
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.2
h5py==3.11.0
huggingface-hub==0.23.0
idna==3.7
imageio==2.34.1
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==3.2.1
keras-tuner==1.4.7
kiwisolver==1.4.5
kt-legacy==1.0.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pudb==2024.1
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
Pygments==2.17.2
pyparsing==3.1.2
PyQt6==6.7.0
PyQt6-Qt6==6.7.1
PyQt6-sip==13.6.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.0
regex==2024.5.15
requests==2.31.0
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
stack-data==0.6.3
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
tf_keras==2.16.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.0
torchtext==0.18.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.2
transformers==4.41.0
triton==2.3.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
urwid==2.6.12
urwid_readline==0.14
wcwidth==0.2.13
Werkzeug==3.0.2
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
